live_trading/requirements.txt
pandas>=1.0.0
numpy>=1.18.0
MetaTrader5>=5.0.0
python-telegram-bot>=13.0.0
psutil>=5.7.0
